File and Mailer service

+++ src/Controller/UserController.php
@@ -4,6 +4,9 @@ namespace App\Controller;
 
 use App\Entity\User;
 use App\Form\UserType;
+use App\Service\FileService;
+use App\Service\Helpers;
+use App\Service\MailerService;
 use Doctrine\Persistence\ManagerRegistry;
 use Exception;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
@@ -12,38 +15,35 @@ use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\HttpFoundation\File\UploadedFile;
 
 use Symfony\Component\Routing\Annotation\Route;
-use Symfony\Component\String\Slugger\SluggerInterface;
 
 class UserController extends AbstractController
 {
     #[Route('/signUp', name: 'app_user_signUp')]
-    public function signUp(Request $request, ManagerRegistry $doctrine, SluggerInterface $slugger): Response
+    public function signUp(Request $request, ManagerRegistry $doctrine, FileService $fileService, MailerService $mailerService): Response
     {
         $newUser = new User();
         $form = $this->createForm(UserType::class, $newUser); 
         
         $form->handleRequest($request);
     
-        if($form->isSubmitted()){
+        if($form->isSubmitted() && $form->isValid()){
             try{
                 /** @var UploadedFile|null $profilePic */
                 $profilePic = $form->get('profilePic')->getData();
 
                 if($profilePic) {
-                    $originalFilename = pathinfo($profilePic->getClientOriginalName(), PATHINFO_FILENAME);
-                    $safeFilename = $slugger->slug($originalFilename);
-                    $newFileName = $safeFilename. '-' . uniqid() . '.' . $profilePic->guessExtension(); 
+                    $localPath =  $fileService->saveProfilePic($profilePic);
+                    $newUser->setProfilePic(pathinfo($localPath, PATHINFO_FILENAME));
                 }
-
-                $localPath = $profilePic->move($this->getParameter('profilePic_directory'), $newFileName)->getPathname();
-                $newUser->setProfilePic($newFileName);
 
                 $manager = $doctrine->getManager();  
                 $manager->persist($newUser);
                 $manager->flush();
                 
                 $this->addFlash('success', $newUser->getFirstName(). ' a été crée avec succès' );
-           
+                
+                $mailerService->sendEmail($newUser->getFirstName());
+
             } catch (Exception $error) {
                 if($profilePic && file_exists($localPath)){
                     unlink($localPath);
@@ -63,5 +63,6 @@ class UserController extends AbstractController
             ]);
         }
     }
+
 
 }
